refactor(racing) : 기능별 패키지 분리
