+++ src/main/java/racingGame/Util.java
@@ -1,4 +1,4 @@
-package RacingGame;
+package racingGame;
 
 import java.util.Arrays;
 import java.util.List;
