+++ src/main/java/racingGame/view/InputView.java
@@ -1,8 +1,8 @@
-package RacingGame;
+package racingGame.view;
 
 import java.util.Scanner;
 
-class InputView {
+public class InputView {
 
     private static Scanner scanner;
 
@@ -10,12 +10,12 @@ class InputView {
         this.scanner = new Scanner(System.in);
     }
 
-    static int countTime() {
+    public static int countTime() {
         System.out.println("시도할 회수는 몇 회 인가요?");
         return scanner.nextInt();
     }
 
-    static String inputNames() {
+    public static String inputNames() {
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
         return scanner.nextLine();
     }
