+++ src/test/java/racingGame/CarTest.java
@@ -1,9 +1,10 @@
-package RacingGame;
+package racingGame;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingGame.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
