+++ src/test/java/racingGame/RacingTest.java
@@ -1,10 +1,11 @@
-package RacingGame;
+package racingGame;
 
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingGame.domain.Racing;
 
 import java.util.List;
