+++ src/test/java/racingGame/UtilTest.java
@@ -1,4 +1,4 @@
-package RacingGame;
+package racingGame;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
